+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add translation directory for installation
 - Add captcha for blog module (https://github.com/PierreRambaud/GotCms/issues/25)
 - Refactoring flash messenger (https://github.com/PierreRambaud/GotCms/issues/26)
+- Backup module failed with pgsql (https://github.com/PierreRambaud/GotCms/issues/27)
 
 ## 0.1.4 (12 Feb 2013):
 - Bug with HTTP 404 (undefined variable view)